routine commit

+++ setup_client.sh
@@ -2,14 +2,30 @@
 # Use: send-setup.sh [optional: container name] [optional: remote:container_id]
 lxcname=${1:-"ocelot"}
 container=${2:-"ubuntu:n"}
-lxc launch $container $lxcname
-wait $!
-LOCATIONNAME="/var/snap/lxd/common/lxd/containers/${lxcname}/rootfs/home/ubuntu"
-KEYFILELOCATION="/home/d/.ssh/container-${lxcname}"
-rm -rf $LOCATIONNAME/container-setup-script
-git clone --depth 1 https://github.com/davidthegardens/container-setup-script.git $LOCATIONNAME/container-setup-script
-sudo touch $LOCATIONNAME/.bashrc
-echo 'sudo /home/ubuntu/container-setup-script/setup_container.sh; exit' | cat - $LOCATIONNAME/.bashrc >temp && mv temp $LOCATIONNAME/.bashrc
+
+IFS=','
+hasBuiltOcelot=false
+while read -r first_column rest_of_line; do
+  # Check if the first column matches the search string
+  if [[ "$first_column" == "ocelot-2.0.0" ]]; then
+    hasBuiltOcelot=true
+    exit 0
+  fi
+done <"$(lxc image list -f csv)"
+
+if [hasBuiltOcelot]; then
+  lxc launch "ocelot-2.0.0" $lxcname
+else
+  lxc launch $container $lxcname
+  wait $!
+  LOCATIONNAME="/var/snap/lxd/common/lxd/containers/${lxcname}/rootfs/home/ubuntu"
+  KEYFILELOCATION="/home/d/.ssh/container-${lxcname}"
+  rm -rf $LOCATIONNAME/container-setup-script
+  git clone --depth 1 https://github.com/davidthegardens/container-setup-script.git $LOCATIONNAME/container-setup-script
+  sudo touch $LOCATIONNAME/.bashrc
+  echo 'sudo /home/ubuntu/container-setup-script/setup_container.sh; exit' | cat - $LOCATIONNAME/.bashrc >temp && mv temp $LOCATIONNAME/.bashrc
+
+fi
 
 sudo -u d ssh-keygen -t ed25519 -f $KEYFILELOCATION
 touch "${LOCATIONNAME}/.ssh/authorized_keys"
